API: Fix getting empty playback state
This commit implements support for Spotify's get playback state endpoint returning 204 No Content. The documentation states that if there is no valid playback state to query, this endpoint will return the mentioned status code.

When we see this, simply return null instead of attempting to de-serialize the response. We have also adjusted the call-sites to handle null appropriately.