API: Offload processing of Game State processing to queue function
This is a performance optimization to ensure that the HTTP triggers return as fast as possible. We do not perform any API calls within them, and instead post a message to the storage queue named `game-state`. Then another function is able to pick up items off that queue and process them without holding up the HTTP clients.